revise type from integer to bigdecimal for story5

+++ kbazaar/src/main/java/com/kampus/kbazaar/cart/CartRequest.java
@@ -2,5 +2,5 @@ package com.kampus.kbazaar.cart;
 
 import java.math.BigDecimal;
 
-public record CartRequest(String sku, String name, BigDecimal price, Integer quantity, String promotionCodes, BigDecimal discount) {
+public record CartRequest(String sku, String name, BigDecimal price, BigDecimal quantity, String promotionCodes, BigDecimal discount) {
 }
